+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Attendance counter: include jQuery Countdown library
 * Attendance counter: minutes until next refresh of the cron event
 * Attendance counter: refresh button to force run of cron event
+* Check Dialog Insight API connection on saving settings
 
 ### 1.0.1: March 1st, 2021
 * Encode token in URL